Refactor: DataGrid Component
